Replaced short usage with byte at necessary places

+++ Applet/AndroidSEProvider/src/com/android/javacard/keymaster/KMAndroidSEApplet.java
@@ -20,10 +20,7 @@ import org.globalplatform.upgrade.OnUpgradeListener;
 import org.globalplatform.upgrade.UpgradeManager;
 
 import com.android.javacard.seprovider.KMAndroidSEProvider;
-import com.android.javacard.seprovider.KMDeviceUniqueKeyPair;
-import com.android.javacard.seprovider.KMError;
 import com.android.javacard.seprovider.KMException;
-import com.android.javacard.seprovider.KMType;
 
 import javacard.framework.APDU;
 import javacard.framework.ISO7816;
@@ -65,7 +62,7 @@ public class KMAndroidSEApplet extends KMKeymasterApplet implements OnUpgradeLis
   private static final byte INS_KEYMINT_PROVIDER_APDU_END = 0x1F;
   public static final byte BOOT_KEY_MAX_SIZE = 32;
   public static final byte BOOT_HASH_MAX_SIZE = 32;
-  public static final short SHARED_SECRET_KEY_SIZE = 32;
+  public static final byte SHARED_SECRET_KEY_SIZE = 32;
 
   // Package version.
   protected short packageVersion;
